injected skill groups and skills (now correctly as singletons).
Moreover: I like the mechanism for creating the skillgroups based on all available skilltypes. The skillgroup constructor expects to receive all available skilltypes and selects the types it knows of belonging to it. This way the skilltypes do not need to know anything about which group they belong to. also it is important that the skills are injected to the groups, because the skills are supposed to be singletons.